Moved and Edited Files
Should now be better organized.
